+++ src/ansys/aedt/core/__init__.py
@@ -88,25 +88,6 @@ from ansys.aedt.core.edb import Edb  # nosec
 from ansys.aedt.core.edb import Siwave  # nosec
 from ansys.aedt.core.generic import constants
 import ansys.aedt.core.generic.data_handlers as data_handler
-from ansys.aedt.core.generic.design_types import Circuit
-from ansys.aedt.core.generic.design_types import CircuitNetlist
-from ansys.aedt.core.generic.design_types import Desktop
-from ansys.aedt.core.generic.design_types import Emit
-from ansys.aedt.core.generic.design_types import FilterSolutions
-from ansys.aedt.core.generic.design_types import Hfss
-from ansys.aedt.core.generic.design_types import Hfss3dLayout
-from ansys.aedt.core.generic.design_types import Icepak
-from ansys.aedt.core.generic.design_types import Maxwell2d
-from ansys.aedt.core.generic.design_types import Maxwell3d
-from ansys.aedt.core.generic.design_types import MaxwellCircuit
-from ansys.aedt.core.generic.design_types import Mechanical
-from ansys.aedt.core.generic.design_types import Q2d
-from ansys.aedt.core.generic.design_types import Q3d
-from ansys.aedt.core.generic.design_types import Rmxprt
-from ansys.aedt.core.generic.design_types import Simplorer
-from ansys.aedt.core.generic.design_types import TwinBuilder
-from ansys.aedt.core.generic.design_types import get_pyaedt_app
-from ansys.aedt.core.generic.design_types import launch_desktop
 from ansys.aedt.core.generic.file_utils import generate_unique_folder_name
 from ansys.aedt.core.generic.file_utils import generate_unique_name
 from ansys.aedt.core.generic.file_utils import generate_unique_project_name
@@ -118,3 +99,22 @@ from ansys.aedt.core.generic.general_methods import is_windows
 from ansys.aedt.core.generic.general_methods import online_help
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
 from ansys.aedt.core.generic.numbers import Quantity
+from ansys.aedt.core.internal.design_types import Circuit
+from ansys.aedt.core.internal.design_types import CircuitNetlist
+from ansys.aedt.core.internal.design_types import Desktop
+from ansys.aedt.core.internal.design_types import Emit
+from ansys.aedt.core.internal.design_types import FilterSolutions
+from ansys.aedt.core.internal.design_types import Hfss
+from ansys.aedt.core.internal.design_types import Hfss3dLayout
+from ansys.aedt.core.internal.design_types import Icepak
+from ansys.aedt.core.internal.design_types import Maxwell2d
+from ansys.aedt.core.internal.design_types import Maxwell3d
+from ansys.aedt.core.internal.design_types import MaxwellCircuit
+from ansys.aedt.core.internal.design_types import Mechanical
+from ansys.aedt.core.internal.design_types import Q2d
+from ansys.aedt.core.internal.design_types import Q3d
+from ansys.aedt.core.internal.design_types import Rmxprt
+from ansys.aedt.core.internal.design_types import Simplorer
+from ansys.aedt.core.internal.design_types import TwinBuilder
+from ansys.aedt.core.internal.design_types import get_pyaedt_app
+from ansys.aedt.core.internal.design_types import launch_desktop
