+++ src/ansys/aedt/core/visualization/post/__init__.py
@@ -34,7 +34,7 @@ def post_processor(app=None, project=None, design=None, version=None):
         PostProcessor object.
     """
     if not app:
-        from ansys.aedt.core.generic.design_types import get_pyaedt_app
+        from ansys.aedt.core.internal.design_types import get_pyaedt_app
         from ansys.aedt.core.internal.desktop_sessions import _desktop_sessions
 
         if not _desktop_sessions:
